--- tests/unit/train/rl/test_chunked_logprobs.py
@@ -1,72 +1,0 @@
-import torch
-
-from prime_rl.trainer.models.layers.lm_head import FusedOutputLinear
-from prime_rl.trainer.rl.loss import compute_entropy
-
-
-def _baseline_logprobs_and_entropy(
-    hidden: torch.Tensor, weight: torch.Tensor, labels: torch.Tensor, *, temperature: float
-) -> tuple[torch.Tensor, torch.Tensor]:
-    logits = hidden @ weight.t()
-    logits = logits / float(temperature)
-    logp = torch.log_softmax(logits, dim=-1).gather(dim=-1, index=labels.unsqueeze(-1)).squeeze(-1)
-    ent = compute_entropy(logits)
-    return logp, ent
-
-
-def test_fused_lm_head_matches_full_logits_forward_and_backward_cpu():
-    torch.manual_seed(0)
-    b, s, h, v = 2, 4, 8, 37
-    temperature = 1.7
-    chunk_size = 11
-
-    hidden0 = torch.randn(b, s, h, dtype=torch.float32, requires_grad=True)
-    labels = torch.randint(0, v, (b, s), dtype=torch.long)
-    weight0 = torch.randn(v, h, dtype=torch.float32, requires_grad=True)
-
-    # Baseline
-    logp0, ent0 = _baseline_logprobs_and_entropy(hidden0, weight0, labels, temperature=temperature)
-    loss0 = logp0.sum()
-    loss0.backward()
-    grad_hidden0 = hidden0.grad.detach().clone()
-    grad_weight0 = weight0.grad.detach().clone()
-
-    # Fused
-    hidden1 = hidden0.detach().clone().requires_grad_(True)
-    weight1 = weight0.detach().clone().requires_grad_(True)
-    lm = FusedOutputLinear(in_features=h, out_features=v, chunk_size=chunk_size)
-    lm.weight = torch.nn.Parameter(weight1)
-
-    out = lm(hidden1, labels, temperature=temperature)
-    assert out.logits is None
-    assert out.logprobs is not None
-    assert out.entropy is not None
-
-    loss1 = out.logprobs.sum()
-    loss1.backward()
-    grad_hidden1 = hidden1.grad.detach().clone()
-    grad_weight1 = lm.weight.grad.detach().clone()
-
-    torch.testing.assert_close(out.logprobs, logp0, rtol=0, atol=1e-5)
-    torch.testing.assert_close(out.entropy, ent0, rtol=0, atol=1e-5)
-    torch.testing.assert_close(grad_hidden1, grad_hidden0, rtol=0, atol=1e-5)
-    torch.testing.assert_close(grad_weight1, grad_weight0, rtol=0, atol=1e-5)
-
-
-def test_fused_lm_head_labels_none_returns_logits():
-    torch.manual_seed(0)
-    b, s, h, v = 2, 3, 4, 9
-
-    hidden = torch.randn(b, s, h, dtype=torch.float32)
-    weight = torch.randn(v, h, dtype=torch.float32)
-
-    lm = FusedOutputLinear(in_features=h, out_features=v, chunk_size=5)
-    lm.weight = torch.nn.Parameter(weight)
-
-    out = lm(hidden, labels=None, temperature=1.0)
-    assert out.logits is not None
-    assert out.logprobs is None
-    assert out.entropy is None
-
-    logits_ref = hidden @ weight.t()
-    torch.testing.assert_close(out.logits, logits_ref, rtol=0, atol=1e-6)
